+++ function/iQIYI-DailyBonus/iqiyi.js
@@ -1,133 +1,89 @@
-//V2ex-Auto-Sign
-const $ = new Env('V2ex自动签到');
-const notify = $.isNode() ? require('./sendNotify') : '';
-const cookie = process.env.V2EXCK
+// version v0.0.1
+// create by BlueSkyClouds
+// detail url: https://github.com/BlueskyClouds/My-Actions
+const exec = require('child_process').execSync
+const fs = require('fs')
+const download = require('download')
+
+const $ = new Env('爱奇艺会员签到');
+const notify = $.isNode() ? require('../sendNotify') : '';
+// 公共变量
+const KEY = process.env.iQIYI_COOKIE
 const SEND_KEY = process.env.SEND_KEY
-const axios = require("axios")
 const UTC8 = new Date().getTime() + new Date().getTimezoneOffset()*60*1000 + 8*60*60*1000;
 
-once = null;
-ckstatus = 1;
-signstatus = 0;
-notice = timeFormat(UTC8) + "\n";
-
-const header = {
-    headers: {
-        Referer: "https://www.v2ex.com/mission",
-        Host: "www.v2ex.com",
-        "user-agent": "Mozilla/5.0 (Linux; Android 10; Redmi K30) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/83.0.4103.83 Mobile Safari/537.36",
-        cookie: `'${cookie}'`,
-    },
-};
-
-//获取once检查是否已签到
-function check() {
-    return new Promise(async (resolve) => {
-        try {
-            let url = "https://www.v2ex.com/mission/daily";
-            let res = await axios.get(url, header);
-            reg1 = /需要先登录/;
-            if (reg1.test(res.data)) {
-                console.log("cookie失效");
-                ckstatus = 0;
-                notice += "cookie失效";
-                if(SEND_KEY){
-                    notify.sendNotify("V2ex自动签到", notice);
-                    return;
-                }
-            } else {
-                reg = /每日登录奖励已领取/;
-                if (reg.test(res.data)) {
-                    notice += "今天已经签到过啦\n";
-                    signstatus = 1;
-                } else {
-                    reg = /redeem\?once=(.*?)'/;
-                    once = res.data.match(reg)[1];
-                    console.log(`获取成功 once:${once}`);
-                }
-            }
-        } catch (err) {
-            console.log(err);
-        }
-        resolve();
-    });
+async function downFile () {
+    const url = 'https://raw.githubusercontent.com/NobyDa/Script/master/iQIYI-DailyBonus/iQIYI.js'
+    await download(url, './')
 }
 
-//每日签到
-function daily() {
-    return new Promise(async (resolve) => {
-        try {
-            let url = `https://www.v2ex.com/mission/daily/redeem?once=${once}`;
-            let res = await axios.get(url, header);
-            reg = /已成功领取每日登录奖励/;
-            if (reg.test(res.data)) {
-                notice += "签到成功\n";
-                signstatus = 1;
-            } else {
-                notice += "签到失败Cookie疑似失效\n";
-                if(SEND_KEY){
-                    notify.sendNotify("V2ex自动签到", notice);
-                    return;
-                }
-            }
-        } catch (err) {
-            console.log(err);
-        }
-        resolve();
-    });
+async function changeFiele () {
+    let content = await fs.readFileSync('./iQIYI.js', 'utf8')
+    content = content.replace(/var cookie = ''/, `var cookie = '${KEY}'`)
+    content = content.replace(/var out = 0/, `var out = 5000`)
+    content = content.replace("${msg ? Details : `response:\\n${data}`}`)", "${msg ? Details : `response:\\n${data}`}`)" + "\n" + "            console.log(data)" + "\n" + "            s = s + 500" + "\n" + "            if(s <= 4500){" + "\n" + "                await Lottery(s)" + "\n" + "            }")
+    await fs.writeFileSync( './iQIYI.js', content, 'utf8')
 }
 
-//查询余额
-function balance() {
-    return new Promise(async (resolve) => {
-        try {
-            let url = "https://www.v2ex.com/balance";
-            let res = await axios.get(url, header);
-            reg = /\d+?\s的每日登录奖励\s\d+\s铜币/;
-            console.log(res.data.match(reg)[0]);
-            notice += res.data.match(reg)[0];
-        } catch (err) {
-            console.log(err);
-        }
-        resolve();
-    });
+async function deleteFile(path) {
+    // 查看文件result.txt是  否存在,如果存在,先删除
+    const fileExists = await fs.existsSync(path);
+    // console.log('fileExists', fileExists);
+    if (fileExists) {
+        const unlinkRes = await fs.unlinkSync(path);
+        // console.log('unlinkRes', unlinkRes)
+    }
 }
 
-function sign() {
-    return new Promise(async (resolve) => {
-        try {
-
-            if (!cookie) {
-                console.log("你的cookie呢！！！");
-                return;
-            }
-            await check();
-            if (once && signstatus === 0) {
-                await daily();
-                await balance();
-                if (signstatus === 0) {
-                    console.log("签到失败Cookie疑似失效")
-                }
-            }
-            console.log(notice);
-            notify.sendNotify("V2ex自动签到", notice);
-        } catch (err) {
-            console.log(err);
+async function start() {
+    if (!KEY) {
+        console.log('请填写 key 后在继续')
+        return
+    }
+    // 下载最新代码
+    await downFile();
+    console.log('下载代码完毕')
+    // 替换变量
+    await changeFiele();
+    console.log('替换变量完毕')
+    // 执行
+    await exec("node iQIYI.js >> result.txt");
+    console.log('执行完毕')
+    const path = "./result.txt";
+    let content = "";
+    if (fs.existsSync(path)) {
+        content = fs.readFileSync(path, "utf8");
+    }
+    if (content.includes("今日已签到")) {
+        //重复签到,不推送仅输出，因为每天会签到两次防止抽奖失败.
+        console.log("爱奇艺签到-" + content + '\n重复签到，取消推送')
+    }else if(SEND_KEY) {
+        if (content.includes("Cookie")) {
+            await notify.sendNotify("爱奇艺签到-" + timeFormat(UTC8), content);
+            console.log("爱奇艺签到-" + content)
+        }else{
+            console.log("爱奇艺签到-" + content)
         }
-        resolve();
-    });
+    }else{
+        await notify.sendNotify("爱奇艺签到-" + timeFormat(UTC8), content);
+        console.log("爱奇艺签到-" + content)
+    }
+
+    //运行完成后，删除下载的文件
+    console.log('运行完成后，删除下载的文件\n')
+    await deleteFile(path);
 }
 
-sign();
+start()
 
 function timeFormat(time) {
-    let date;
-    if (time) {
-        date = new Date(time)
-    } else {
-        date = new Date();
-    }
-    return date.getFullYear() + '年' + ((date.getMonth() + 1) >= 10 ? (date.getMonth() + 1) : '0' + (date.getMonth() + 1)) + '月' + (date.getDate() >= 10 ? date.getDate() : '0' + date.getDate()) + '日';
+  let date;
+  if (time) {
+    date = new Date(time)
+  } else {
+    date = new Date();
+  }
+  return date.getFullYear() + '年' + ((date.getMonth() + 1) >= 10 ? (date.getMonth() + 1) : '0' + (date.getMonth() + 1)) + '月' + (date.getDate() >= 10 ? date.getDate() : '0' + date.getDate()) + '日';
 }
 // prettier-ignore
 function Env(t,e){class s{constructor(t){this.env=t}send(t,e="GET"){t="string"==typeof t?{url:t}:t;let s=this.get;return"POST"===e&&(s=this.post),new Promise((e,i)=>{s.call(this,t,(t,s,r)=>{t?i(t):e(s)})})}get(t){return this.send.call(this.env,t)}post(t){return this.send.call(this.env,t,"POST")}}return new class{constructor(t,e){this.name=t,this.http=new s(this),this.data=null,this.dataFile="box.dat",this.logs=[],this.isMute=!1,this.isNeedRewrite=!1,this.logSeparator="\n",this.startTime=(new Date).getTime(),Object.assign(this,e),this.log("",`\ud83d\udd14${this.name}, \u5f00\u59cb!`)}isNode(){return"undefined"!=typeof module&&!!module.exports}isQuanX(){return"undefined"!=typeof $task}isSurge(){return"undefined"!=typeof $httpClient&&"undefined"==typeof $loon}isLoon(){return"undefined"!=typeof $loon}isShadowrocket(){return"undefined"!=typeof $rocket}toObj(t,e=null){try{return JSON.parse(t)}catch{return e}}toStr(t,e=null){try{return JSON.stringify(t)}catch{return e}}getjson(t,e){let s=e;const i=this.getdata(t);if(i)try{s=JSON.parse(this.getdata(t))}catch{}return s}setjson(t,e){try{return this.setdata(JSON.stringify(t),e)}catch{return!1}}getScript(t){return new Promise(e=>{this.get({url:t},(t,s,i)=>e(i))})}runScript(t,e){return new Promise(s=>{let i=this.getdata("@chavy_boxjs_userCfgs.httpapi");i=i?i.replace(/\n/g,"").trim():i;let r=this.getdata("@chavy_boxjs_userCfgs.httpapi_timeout");r=r?1*r:20,r=e&&e.timeout?e.timeout:r;const[o,h]=i.split("@"),a={url:`http://${h}/v1/scripting/evaluate`,body:{script_text:t,mock_type:"cron",timeout:r},headers:{"X-Key":o,Accept:"*/*"}};this.post(a,(t,e,i)=>s(i))}).catch(t=>this.logErr(t))}loaddata(){if(!this.isNode())return{};{this.fs=this.fs?this.fs:require("fs"),this.path=this.path?this.path:require("path");const t=this.path.resolve(this.dataFile),e=this.path.resolve(process.cwd(),this.dataFile),s=this.fs.existsSync(t),i=!s&&this.fs.existsSync(e);if(!s&&!i)return{};{const i=s?t:e;try{return JSON.parse(this.fs.readFileSync(i))}catch(t){return{}}}}}writedata(){if(this.isNode()){this.fs=this.fs?this.fs:require("fs"),this.path=this.path?this.path:require("path");const t=this.path.resolve(this.dataFile),e=this.path.resolve(process.cwd(),this.dataFile),s=this.fs.existsSync(t),i=!s&&this.fs.existsSync(e),r=JSON.stringify(this.data);s?this.fs.writeFileSync(t,r):i?this.fs.writeFileSync(e,r):this.fs.writeFileSync(t,r)}}lodash_get(t,e,s){const i=e.replace(/\[(\d+)\]/g,".$1").split(".");let r=t;for(const t of i)if(r=Object(r)[t],void 0===r)return s;return r}lodash_set(t,e,s){return Object(t)!==t?t:(Array.isArray(e)||(e=e.toString().match(/[^.[\]]+/g)||[]),e.slice(0,-1).reduce((t,s,i)=>Object(t[s])===t[s]?t[s]:t[s]=Math.abs(e[i+1])>>0==+e[i+1]?[]:{},t)[e[e.length-1]]=s,t)}getdata(t){let e=this.getval(t);if(/^@/.test(t)){const[,s,i]=/^@(.*?)\.(.*?)$/.exec(t),r=s?this.getval(s):"";if(r)try{const t=JSON.parse(r);e=t?this.lodash_get(t,i,""):e}catch(t){e=""}}return e}setdata(t,e){let s=!1;if(/^@/.test(e)){const[,i,r]=/^@(.*?)\.(.*?)$/.exec(e),o=this.getval(i),h=i?"null"===o?null:o||"{}":"{}";try{const e=JSON.parse(h);this.lodash_set(e,r,t),s=this.setval(JSON.stringify(e),i)}catch(e){const o={};this.lodash_set(o,r,t),s=this.setval(JSON.stringify(o),i)}}else s=this.setval(t,e);return s}getval(t){return this.isSurge()||this.isLoon()?$persistentStore.read(t):this.isQuanX()?$prefs.valueForKey(t):this.isNode()?(this.data=this.loaddata(),this.data[t]):this.data&&this.data[t]||null}setval(t,e){return this.isSurge()||this.isLoon()?$persistentStore.write(t,e):this.isQuanX()?$prefs.setValueForKey(t,e):this.isNode()?(this.data=this.loaddata(),this.data[e]=t,this.writedata(),!0):this.data&&this.data[e]||null}initGotEnv(t){this.got=this.got?this.got:require("got"),this.cktough=this.cktough?this.cktough:require("tough-cookie"),this.ckjar=this.ckjar?this.ckjar:new this.cktough.CookieJar,t&&(t.headers=t.headers?t.headers:{},void 0===t.headers.Cookie&&void 0===t.cookieJar&&(t.cookieJar=this.ckjar))}get(t,e=(()=>{})){t.headers&&(delete t.headers["Content-Type"],delete t.headers["Content-Length"]),this.isSurge()||this.isLoon()?(this.isSurge()&&this.isNeedRewrite&&(t.headers=t.headers||{},Object.assign(t.headers,{"X-Surge-Skip-Scripting":!1})),$httpClient.get(t,(t,s,i)=>{!t&&s&&(s.body=i,s.statusCode=s.status),e(t,s,i)})):this.isQuanX()?(this.isNeedRewrite&&(t.opts=t.opts||{},Object.assign(t.opts,{hints:!1})),$task.fetch(t).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>e(t))):this.isNode()&&(this.initGotEnv(t),this.got(t).on("redirect",(t,e)=>{try{if(t.headers["set-cookie"]){const s=t.headers["set-cookie"].map(this.cktough.Cookie.parse).toString();s&&this.ckjar.setCookieSync(s,null),e.cookieJar=this.ckjar}}catch(t){this.logErr(t)}}).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>{const{message:s,response:i}=t;e(s,i,i&&i.body)}))}post(t,e=(()=>{})){const s=t.method?t.method.toLocaleLowerCase():"post";if(t.body&&t.headers&&!t.headers["Content-Type"]&&(t.headers["Content-Type"]="application/x-www-form-urlencoded"),t.headers&&delete t.headers["Content-Length"],this.isSurge()||this.isLoon())this.isSurge()&&this.isNeedRewrite&&(t.headers=t.headers||{},Object.assign(t.headers,{"X-Surge-Skip-Scripting":!1})),$httpClient[s](t,(t,s,i)=>{!t&&s&&(s.body=i,s.statusCode=s.status),e(t,s,i)});else if(this.isQuanX())t.method=s,this.isNeedRewrite&&(t.opts=t.opts||{},Object.assign(t.opts,{hints:!1})),$task.fetch(t).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>e(t));else if(this.isNode()){this.initGotEnv(t);const{url:i,...r}=t;this.got[s](i,r).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>{const{message:s,response:i}=t;e(s,i,i&&i.body)})}}time(t,e=null){const s=e?new Date(e):new Date;let i={"M+":s.getMonth()+1,"d+":s.getDate(),"H+":s.getHours(),"m+":s.getMinutes(),"s+":s.getSeconds(),"q+":Math.floor((s.getMonth()+3)/3),S:s.getMilliseconds()};/(y+)/.test(t)&&(t=t.replace(RegExp.$1,(s.getFullYear()+"").substr(4-RegExp.$1.length)));for(let e in i)new RegExp("("+e+")").test(t)&&(t=t.replace(RegExp.$1,1==RegExp.$1.length?i[e]:("00"+i[e]).substr((""+i[e]).length)));return t}msg(e=t,s="",i="",r){const o=t=>{if(!t)return t;if("string"==typeof t)return this.isLoon()?t:this.isQuanX()?{"open-url":t}:this.isSurge()?{url:t}:void 0;if("object"==typeof t){if(this.isLoon()){let e=t.openUrl||t.url||t["open-url"],s=t.mediaUrl||t["media-url"];return{openUrl:e,mediaUrl:s}}if(this.isQuanX()){let e=t["open-url"]||t.url||t.openUrl,s=t["media-url"]||t.mediaUrl;return{"open-url":e,"media-url":s}}if(this.isSurge()){let e=t.url||t.openUrl||t["open-url"];return{url:e}}}};if(this.isMute||(this.isSurge()||this.isLoon()?$notification.post(e,s,i,o(r)):this.isQuanX()&&$notify(e,s,i,o(r))),!this.isMuteLog){let t=["","==============\ud83d\udce3\u7cfb\u7edf\u901a\u77e5\ud83d\udce3=============="];t.push(e),s&&t.push(s),i&&t.push(i),console.log(t.join("\n")),this.logs=this.logs.concat(t)}}log(...t){t.length>0&&(this.logs=[...this.logs,...t]),console.log(t.join(this.logSeparator))}logErr(t,e){const s=!this.isSurge()&&!this.isQuanX()&&!this.isLoon();s?this.log("",`\u2757\ufe0f${this.name}, \u9519\u8bef!`,t.stack):this.log("",`\u2757\ufe0f${this.name}, \u9519\u8bef!`,t)}wait(t){return new Promise(e=>setTimeout(e,t))}done(t={}){const e=(new Date).getTime(),s=(e-this.startTime)/1e3;this.log("",`\ud83d\udd14${this.name}, \u7ed3\u675f! \ud83d\udd5b ${s} \u79d2`),this.log(),(this.isSurge()||this.isQuanX()||this.isLoon())&&$done(t)}}(t,e)}
